requirements.txt
asgiref==3.3.1
cffi==1.14.4
cryptography==3.3.1
Django==3.1.5
django-classy-tags==2.0.0
django-taggit==1.3.0
django-taggit-templatetags2==1.6.1
django-widget-tweaks==1.4.8
ndg-httpsclient==0.5.1
Pillow==8.1.0
pyasn1==0.4.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
